+++ src/pages/liquidations/index.tsx
@@ -1,33 +1,16 @@
 /* eslint-disable no-unused-vars*/
 // eslint sucks at types
 import { NextPage, GetServerSideProps } from 'next'
-import { useRouter } from 'next/router'
-import { PlusCircle } from 'react-feather'
 import useSWR from 'swr'
-import { fetcher, arrayFetcher, retrySWR } from '~/utils/useSWR'
-import { Select as AriaSelect, SelectItem, SelectPopover } from 'ariakit/select'
-
-import { ButtonDark } from '~/components/ButtonStyled'
+import { fetcher } from '~/utils/useSWR'
 import Layout from '~/layout'
 // import { revalidate } from '~/api'
 
-// import { liqs } from '../../components/LiquidationsPage'
 import { Header } from '~/Theme'
 import { ProtocolsChainsSearch } from '~/components/Search'
-import React, { useEffect, useMemo } from 'react'
 import { ChartData, useLiquidationsState } from '~/utils/liquidations'
-import { BreakpointPanelNoBorder } from '~/components'
-import HeadHelp from '~/components/HeadHelp'
-import { ChartState } from '../../components/LiquidationsPage/Chart'
-import { LiquidationsContainer } from '../../components/LiquidationsPage/Main'
-
-import { queryTypes, useQueryState } from 'next-usequerystate'
-
-export const defaultChartState: ChartState = {
-	asset: 'ETH',
-	aggregateBy: 'protocol',
-	filters: ['all'] // comma separated list of chains or protocols
-}
+import { ChartState } from '../../components/LiquidationsPage/utils'
+import { LiquidationsContainer } from '../../components/LiquidationsPage/LiquidationsContainer'
 
 const LiquidationsPage: NextPage = () => {
 	const { asset, aggregateBy, filters } = useLiquidationsState()
@@ -35,7 +18,7 @@ const LiquidationsPage: NextPage = () => {
 	const { data } = useSWR<ChartData>(
 		// TODO: implement the full api
 		`http://localhost:3000/api/mock-liquidations/?symbol=${asset}&aggregateBy=${aggregateBy}${
-			!filters ? '' : `&filters=${filters}`
+			filters.includes('all') ? '' : `&filters=${filters.join(',')}`
 		}`,
 		fetcher
 	)
@@ -47,14 +30,7 @@ const LiquidationsPage: NextPage = () => {
 			/>
 
 			<Header>Liquidation Levels in DeFi 💦</Header>
-			{data && (
-				<LiquidationsContainer
-					chartState={{ asset, aggregateBy, filters }}
-					chartData={data}
-					uid={`liquidations-chart-${asset}`}
-					aggregateBy={aggregateBy}
-				/>
-			)}
+			{data && <LiquidationsContainer chartState={{ asset, aggregateBy, filters }} chartData={data} />}
 		</Layout>
 	)
 }
